requirements.txt
langchain==0.3.18
langgraph==0.2.70
langchain-openai==0.3.4
python-dotenv==1.0.1
streamlit==1.42.0